// FE/src/api/generated/models/paginatedProductsResponseDataItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * awr-store-backend-project
 * OpenAPI spec version: 1.0.0
 */

export type PaginatedProductsResponseDataItem = {
  /**
   * Unique product id
   * @minimum 0
   * @exclusiveMinimum
   */
  id: number;
  /**
   * Product name
   * @minLength 1
   */
  name: string;
  /** Product description */
  description: string;
  /**
   * Product price
   * @minimum 0
   */
  price: number;
  /**
   * Product quantity available for fulfillment
   * @minimum 0
   */
  availableCount: number;
  /** Product creation timestamp */
  createdAt: unknown;
  /** Product last updated timestamp */
  updatedAt: unknown;
};